Move all backend types to the interface when possible.
