+++ luminance/src/backend/buffer.rs
@@ -2,8 +2,7 @@
 //!
 //! This interface defines the low-level API buffers must implement to be usable.
 
-use std::fmt;
-
+use crate::buffer::BufferError;
 use crate::linear::{M22, M33, M44};
 
 pub unsafe trait BufferBase {
@@ -45,66 +44,6 @@ pub unsafe trait Buffer<T>: BufferBase {
   unsafe fn clear(buffer: &mut Self::BufferRepr, x: T) -> Result<(), BufferError>
   where
     T: Copy;
-}
-
-/// Buffer errors.
-#[derive(Debug, Eq, PartialEq)]
-pub enum BufferError {
-  /// Overflow when setting a value with a specific index.
-  ///
-  /// Contains the index and the size of the buffer.
-  Overflow { index: usize, buffer_len: usize },
-
-  /// Too few values were passed to fill a buffer.
-  ///
-  /// Contains the number of passed value and the size of the buffer.
-  TooFewValues {
-    provided_len: usize,
-    buffer_len: usize,
-  },
-
-  /// Too many values were passed to fill a buffer.
-  ///
-  /// Contains the number of passed value and the size of the buffer.
-  TooManyValues {
-    provided_len: usize,
-    buffer_len: usize,
-  },
-
-  /// Mapping the buffer failed.
-  MapFailed,
-}
-
-impl fmt::Display for BufferError {
-  fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-    match *self {
-      BufferError::Overflow { index, buffer_len } => write!(
-        f,
-        "buffer overflow (index = {}, size = {})",
-        index, buffer_len
-      ),
-
-      BufferError::TooFewValues {
-        provided_len,
-        buffer_len,
-      } => write!(
-        f,
-        "too few values passed to the buffer (nb = {}, size = {})",
-        provided_len, buffer_len
-      ),
-
-      BufferError::TooManyValues {
-        provided_len,
-        buffer_len,
-      } => write!(
-        f,
-        "too many values passed to the buffer (nb = {}, size = {})",
-        provided_len, buffer_len
-      ),
-
-      BufferError::MapFailed => write!(f, "buffer mapping failed"),
-    }
-  }
 }
 
 pub unsafe trait BufferSlice<T>: Buffer<T> {
