+++ luminance/src/pipeline.rs
@@ -2,18 +2,146 @@ use crate::backend::color_slot::ColorSlot;
 use crate::backend::depth_slot::DepthSlot;
 use crate::backend::framebuffer::Framebuffer as FramebufferBackend;
 use crate::backend::pipeline::{
-  Pipeline as PipelineBackend, PipelineBase, PipelineBuffer, PipelineError, PipelineState,
-  PipelineTexture,
+  Pipeline as PipelineBackend, PipelineBase, PipelineBuffer, PipelineTexture,
 };
-use crate::backend::texture::{Dimensionable, Layerable};
 use crate::buffer::Buffer;
 use crate::context::GraphicsContext;
 use crate::framebuffer::Framebuffer;
 use crate::pixel::Pixel;
 use crate::shading_gate::ShadingGate;
 use crate::texture::Texture;
+use crate::texture::{Dimensionable, Layerable};
 
+use std::fmt;
 use std::marker::PhantomData;
+
+#[derive(Debug)]
+pub enum PipelineError {}
+
+impl fmt::Display for PipelineError {
+  fn fmt(&self, _: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+    Ok(())
+  }
+}
+
+/// The viewport being part of the [`PipelineState`].
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub enum Viewport {
+  /// The whole viewport is used. The position and dimension of the viewport rectangle are
+  /// extracted from the framebuffer.
+  Whole,
+  /// The viewport is specific and the rectangle area is user-defined.
+  Specific {
+    /// The lower position on the X axis to start the viewport rectangle at.
+    x: u32,
+    /// The lower position on the Y axis to start the viewport rectangle at.
+    y: u32,
+    /// The width of the viewport.
+    width: u32,
+    /// The height of the viewport.
+    height: u32,
+  },
+}
+
+/// Various customization options for pipelines.
+#[derive(Clone, Debug)]
+pub struct PipelineState {
+  pub(crate) clear_color: [f32; 4],
+  pub(crate) clear_color_enabled: bool,
+  pub(crate) clear_depth_enabled: bool,
+  pub(crate) viewport: Viewport,
+  pub(crate) srgb_enabled: bool,
+}
+
+impl Default for PipelineState {
+  /// Default [`PipelineState`]:
+  ///
+  /// - Clear color: `[0, 0, 0, 1]`.
+  /// - Color is always cleared.
+  /// - Depth is always cleared.
+  /// - The viewport uses the whole framebuffer’s.
+  /// - sRGB encoding is disabled.
+  fn default() -> Self {
+    PipelineState {
+      clear_color: [0., 0., 0., 1.],
+      clear_color_enabled: true,
+      clear_depth_enabled: true,
+      viewport: Viewport::Whole,
+      srgb_enabled: false,
+    }
+  }
+}
+
+impl PipelineState {
+  /// Create a default [`PipelineState`].
+  ///
+  /// See the documentation of the [`Default`] for further details.
+  pub fn new() -> Self {
+    Self::default()
+  }
+
+  /// Get the clear color.
+  pub fn clear_color(&self) -> [f32; 4] {
+    self.clear_color
+  }
+
+  /// Set the clear color.
+  pub fn set_clear_color(self, clear_color: [f32; 4]) -> Self {
+    Self {
+      clear_color,
+      ..self
+    }
+  }
+
+  /// Check whether the pipeline’s framebuffer’s color buffers will be cleared.
+  pub fn is_clear_color_enabled(&self) -> bool {
+    self.clear_color_enabled
+  }
+
+  /// Enable clearing color buffers.
+  pub fn enable_clear_color(self, clear_color_enabled: bool) -> Self {
+    Self {
+      clear_color_enabled,
+      ..self
+    }
+  }
+
+  /// Check whether the pipeline’s framebuffer’s depth buffer will be cleared.
+  pub fn is_clear_depth_enabled(&self) -> bool {
+    self.clear_depth_enabled
+  }
+
+  /// Enable clearing depth buffers.
+  pub fn enable_clear_depth(self, clear_depth_enabled: bool) -> Self {
+    Self {
+      clear_depth_enabled,
+      ..self
+    }
+  }
+
+  /// Get the viewport.
+  pub fn viewport(&self) -> Viewport {
+    self.viewport
+  }
+
+  /// Set the viewport.
+  pub fn set_viewport(self, viewport: Viewport) -> Self {
+    Self { viewport, ..self }
+  }
+
+  /// Check whether sRGB linearization is enabled.
+  pub fn is_srgb_enabled(&self) -> bool {
+    self.srgb_enabled
+  }
+
+  /// Enable sRGB linearization.
+  pub fn enable_srgb(self, srgb_enabled: bool) -> Self {
+    Self {
+      srgb_enabled,
+      ..self
+    }
+  }
+}
 
 pub struct Pipeline<'a, B>
 where
