+++ luminance/src/tess.rs
@@ -1,13 +1,190 @@
 //! Tessellation API.
 
+use std::fmt;
 use std::ops::{Range, RangeFrom, RangeFull, RangeInclusive, RangeTo, RangeToInclusive};
 
 use crate::backend::tess::{
-  Mode, Tess as TessBackend, TessBuilder as TessBuilderBackend, TessError, TessIndex, TessMapError,
-  TessSlice as TessSliceBackend,
+  Tess as TessBackend, TessBuilder as TessBuilderBackend, TessSlice as TessSliceBackend,
 };
+use crate::buffer::BufferError;
 use crate::context::GraphicsContext;
 use crate::vertex::Vertex;
+use crate::vertex::VertexDesc;
+
+/// Vertices can be connected via several modes.
+///
+/// Some modes allow for _primitive restart_. Primitive restart is a cool feature that allows to
+/// _break_ the building of a primitive to _start over again_. For instance, when making a curve,
+/// you can imagine gluing segments next to each other. If at some point, you want to start a new
+/// line, you have two choices:
+///
+///   - Either you stop your draw call and make another one.
+///   - Or you just use the _primitive restart_ feature to ask to create another line from scratch.
+///
+/// That feature is encoded with a special _vertex index_. You can setup the value of the _primitive
+/// restart index_ with [`TessBuilder::set_primitive_restart_index`]. Whenever a vertex index is set
+/// to the same value as the _primitive restart index_, the value is not interpreted as a vertex
+/// index but just a marker / hint to start a new primitive.
+#[derive(Copy, Clone, Debug)]
+pub enum Mode {
+  /// A single point.
+  ///
+  /// Points are left unconnected from each other and represent a _point cloud_. This is the typical
+  /// primitive mode you want to do, for instance, particles rendering.
+  Point,
+  /// A line, defined by two points.
+  ///
+  /// Every pair of vertices are connected together to form a straight line.
+  Line,
+  /// A strip line, defined by at least two points and zero or many other ones.
+  ///
+  /// The first two vertices create a line, and every new vertex flowing in the graphics pipeline
+  /// (starting from the third, then) well extend the initial line, making a curve composed of
+  /// several segments.
+  ///
+  /// > This kind of primitive mode allows the usage of _primitive restart_.
+  LineStrip,
+  /// A triangle, defined by three points.
+  Triangle,
+  /// A triangle fan, defined by at least three points and zero or many other ones.
+  ///
+  /// Such a mode is easy to picture: a cooling fan is a circular shape, with blades.
+  /// [`Mode::TriangleFan`] is kind of the same. The first vertex is at the center of the fan, then
+  /// the second vertex creates the first edge of the first triangle. Every time you add a new
+  /// vertex, a triangle is created by taking the first (center) vertex, the very previous vertex
+  /// and the current vertex. By specifying vertices around the center, you actually create a
+  /// fan-like shape.
+  ///
+  /// > This kind of primitive mode allows the usage of _primitive restart_.
+  TriangleFan,
+  /// A triangle strip, defined by at least three points and zero or many other ones.
+  ///
+  /// This mode is a bit different from [`Mode::TriangleFan`]. The first two vertices define the
+  /// first edge of the first triangle. Then, for each new vertex, a new triangle is created by
+  /// taking the very previous vertex and the last to very previous vertex. What it means is that
+  /// every time a triangle is created, the next vertex will share the edge that was created to
+  /// spawn the previous triangle.
+  ///
+  /// This mode is useful to create long ribbons / strips of triangles.
+  ///
+  /// > This kind of primitive mode allows the usage of _primitive restart_.
+  TriangleStrip,
+  /// A general purpose primitive with _n_ vertices, for use in tessellation shaders.
+  /// For example, `Mode::Patch(3)` represents triangle patches, so every three vertices in the
+  /// buffer form a patch.
+  /// If you want to employ tessellation shaders, this is the only primitive mode you can use.
+  Patch(usize),
+}
+
+/// Error that can occur while trying to map GPU tessellation to host code.
+#[derive(Debug, Eq, PartialEq)]
+pub enum TessMapError {
+  /// The CPU mapping failed due to buffer errors.
+  BufferMapError(BufferError),
+  /// Vertex target type is not the same as the one stored in the buffer.
+  VertexTypeMismatch(VertexDesc, VertexDesc),
+  /// Index target type is not the same as the one stored in the buffer.
+  IndexTypeMismatch(TessIndexType, TessIndexType),
+  /// The CPU mapping failed because you cannot map an attributeless tessellation since it doesn’t
+  /// have any vertex attribute.
+  ForbiddenAttributelessMapping,
+  /// The CPU mapping failed because currently, mapping deinterleaved buffers is not supported via
+  /// a single slice.
+  ForbiddenDeinterleavedMapping,
+}
+
+impl fmt::Display for TessMapError {
+  fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+    match *self {
+      TessMapError::BufferMapError(ref e) => write!(f, "cannot map tessellation buffer: {}", e),
+      TessMapError::VertexTypeMismatch(ref a, ref b) => write!(
+        f,
+        "cannot map tessellation: vertex type mismatch between {:?} and {:?}",
+        a, b
+      ),
+      TessMapError::IndexTypeMismatch(ref a, ref b) => write!(
+        f,
+        "cannot map tessellation: index type mismatch between {:?} and {:?}",
+        a, b
+      ),
+      TessMapError::ForbiddenAttributelessMapping => {
+        f.write_str("cannot map an attributeless buffer")
+      }
+      TessMapError::ForbiddenDeinterleavedMapping => {
+        f.write_str("cannot map a deinterleaved buffer as interleaved")
+      }
+    }
+  }
+}
+
+/// Possible errors that might occur when dealing with [`Tess`].
+#[derive(Debug)]
+pub enum TessError {
+  /// Error related to attributeless tessellation and/or render.
+  AttributelessError(String),
+  /// Length incoherency in vertex, index or instance buffers.
+  LengthIncoherency(usize),
+  /// Overflow when accessing underlying buffers.
+  Overflow(usize, usize),
+  /// Internal error ocurring with a buffer.
+  InternalBufferError(BufferError),
+}
+
+impl From<BufferError> for TessError {
+  fn from(e: BufferError) -> Self {
+    TessError::InternalBufferError(e)
+  }
+}
+
+/// Possible tessellation index types.
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
+pub enum TessIndexType {
+  /// 8-bit unsigned integer.
+  U8,
+  /// 16-bit unsigned integer.
+  U16,
+  /// 32-bit unsigned integer.
+  U32,
+}
+
+impl TessIndexType {
+  pub fn bytes(self) -> usize {
+    match self {
+      TessIndexType::U8 => 1,
+      TessIndexType::U16 => 2,
+      TessIndexType::U32 => 4,
+    }
+  }
+}
+
+/// Class of tessellation indexes.
+///
+/// Values which types implement this trait are allowed to be used to index tessellation in *indexed
+/// draw commands*.
+///
+/// You shouldn’t have to worry to much about that trait. Have a look at the current implementors
+/// for an exhaustive list of types you can use.
+///
+/// > Implementing this trait is `unsafe`.
+pub unsafe trait TessIndex {
+  /// Type of the underlying index.
+  ///
+  /// You are limited in which types you can use as indexes. Feel free to have a look at the
+  /// documentation of the [`TessIndexType`] trait for further information.
+  const INDEX_TYPE: TessIndexType;
+}
+
+unsafe impl TessIndex for u8 {
+  const INDEX_TYPE: TessIndexType = TessIndexType::U8;
+}
+
+unsafe impl TessIndex for u16 {
+  const INDEX_TYPE: TessIndexType = TessIndexType::U16;
+}
+
+unsafe impl TessIndex for u32 {
+  const INDEX_TYPE: TessIndexType = TessIndexType::U32;
+}
 
 pub struct TessBuilder<'a, C>
 where
